feat: implement core models (Equipment, Facility, FacilityClass, FacilityClassUser, Session, SessionUser, User) and controllers (Facility, User)

+++ source_code/Source/hokinh-taekwondo/src/main/java/com/hokinhtaekwondo/hokinh_taekwondo/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package com.hokinhtaekwondo.hokinh_taekwondo.repository;
 import com.hokinhtaekwondo.hokinh_taekwondo.model.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article, Long> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
 }
