+++ source_code/Source/hokinh-taekwondo/src/main/java/com/hokinhtaekwondo/hokinh_taekwondo/repository/AwardRepository.java
@@ -3,5 +3,5 @@ package com.hokinhtaekwondo.hokinh_taekwondo.repository;
 import com.hokinhtaekwondo.hokinh_taekwondo.model.Award;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AwardRepository extends JpaRepository<Award, Long> {
+public interface AwardRepository extends JpaRepository<Award, Integer> {
 }
